+++ .gitignore
@@ -28,6 +28,7 @@
 !.vscode/extensions.json
 
 # misc
+.angular
 /.sass-cache
 /connect.lock
 /coverage
