+++ package.json
@@ -11,7 +11,7 @@
     "test": "ng test",
     "lint": "ng lint",
     "tslint": "ng lint",
-    "build-release": "ng build arlas-web-components --prod",
+    "build-release": "ng build arlas-web-components --configuration production",
     "e2e": "ng e2e",
     "analyze": "webpack-bundle-analyzer dist/stats-es2015.json",
     "i18n:init": "ngx-translate-extract --input ./projects/arlas-components/src/lib --output ./projects/arlas-components/assets/i18n/template.json --key-as-default-value --replace --format json",
@@ -47,7 +47,7 @@
     "jszip": "^3.2.1",
     "mapbox-gl": "^1.6.1",
     "moment": "^2.24.0",
-    "ng-lazyload-image": "^7.0.1",
+    "ng-lazyload-image": "^9.0.0",
     "ngeohash": "^0.6.0",
     "ngx-loading": "^1.0.7",
     "rxjs": "~7.4.0",
@@ -55,8 +55,8 @@
     "stream": "0.0.2",
     "turf-extent": "1.0.4",
     "wellknown": "0.5.0",
-    "tslib": "^2.3.0",
-    "zone.js": "^0.11.4"
+    "zone.js": "^0.11.4",
+    "tslib": "^2.3.0"
   },
   "devDependencies": {
     "@angular-devkit/build-angular": "~13.0.4",
@@ -64,20 +64,22 @@
     "@angular/compiler-cli": "~13.0.0",
     "@types/jasmine": "~3.10.0",
     "@types/node": "^12.11.1",
+    "jasmine-core": "~3.10.0",
+    "karma": "~6.3.0",
+    "karma-chrome-launcher": "~3.1.0",
+    "karma-coverage": "~2.0.3",
+    "karma-jasmine": "~4.0.0",
+    "karma-jasmine-html-reporter": "~1.7.0",
     "@biesbjerg/ngx-translate-extract": "7.0.4",
     "@biesbjerg/ngx-translate-extract-marker": "1.0.0",
     "gulp": "^4.0.0",
-    "jasmine-core": "~3.10.0",
-    "karma": "^5.0.1",
-    "karma-chrome-launcher": "~2.2.0",
-    "karma-jasmine": "~1.1.2",
-    "karma-jasmine-html-reporter": "^0.2.2",
     "node": "^12.11.1",
     "ts-node": "~10.4.0",
-    "tsickle": ">=0.29.0",
     "tslint": "~5.16.0",
     "typescript": "~4.4.3",
     "ng-packagr": "^13.0.0",
     "webpack-bundle-analyzer": "^3.6.1"
+
+
   }
 }
