+++ projects/arlas-components/package.json
@@ -2,22 +2,21 @@
   "name": "arlas-web-components",
   "version": "0.0.1",
   "peerDependencies": {
-    "@angular/animations": "~11.2.9",
-    "@angular/cdk": "~11.2.9",
-    "@angular/common": "~11.2.9",
-    "@angular/core": "~11.2.9",
-    "@angular/forms": "~11.2.9",
-    "@angular/material": "~11.2.9",
-    "@angular/platform-browser": "~11.2.9",
-    "@angular/platform-browser-dynamic": "~11.2.9",
-    "@angular/router": "~11.2.9",
+    "@angular/animations": "~13.0.0",
+    "@angular/common": "~13.0.0",
+    "@angular/core": "~13.0.0",
+    "@angular/forms": "~13.0.0",
+    "@angular/material": "~13.0.0",
+    "@angular/platform-browser": "~13.0.0",
+    "@angular/platform-browser-dynamic": "~13.0.0",
+    "@angular/router": "~13.0.0",
     "@ngx-translate/core": "^13.0.0",
     "@ngx-translate/http-loader": "^6.0.0",
     "@types/tinycolor2": "^1.4.0",
     "core-js": "^2.5.4",
     "moment": "^2.24.0",
     "stream": "0.0.2",
-    "rxjs": "^6.6.7",
+    "rxjs": "~7.4.0",
     "zone.js": "^0.11.4"
   },
   "dependencies": {
@@ -27,7 +26,7 @@
     "@turf/helpers": "6.1.4",
     "@types/d3-dsv": "1.0.33",
     "@types/mapbox-gl": "^1.6.3",
-    "arlas-d3": "7.5.0-beta.6",
+    "arlas-d3": "9.0.0-beta.4",
     "d3-dsv": "^1.0.10",
     "geojson-polygon-self-intersections": "1.2.1",
     "geojson-validation": "0.2.1",
@@ -35,12 +34,12 @@
     "jsts": "2.1.0",
     "jszip": "^3.2.1",
     "mapbox-gl": "^1.6.1",
-    "ng-lazyload-image": "^7.0.1",
+    "ng-lazyload-image": "^9.0.0",
     "ngeohash": "^0.6.0",
     "ngx-loading": "^1.0.7",
     "shpjs": "^3.4.3",
     "turf-extent": "1.0.4",
     "wellknown": "0.5.0",
-    "tslib": "^2.0.0"
+    "tslib": "^2.3.0"
   }
 }
