+++ projects/arlas-components/src/lib/components/mapgl/mapgl.component.util.ts
@@ -79,3 +79,31 @@ export class GetCollectionPipe implements PipeTransform {
         return !!layersMap ? !!layersMap.get(value).metadata ? layersMap.get(value).metadata.collection : undefined : undefined;
     }
 }
+
+export interface LegendData {
+    minValue?: string;
+    maxValue?: string;
+    keysColorsMap?: Map<string, string>;
+}
+
+export interface Legend {
+    type?: PROPERTY_SELECTOR_SOURCE;
+    title?: string;
+    minValue?: string;
+    maxValue?: string;
+    fixValue?: string | number;
+    interpolatedValues?: Array<string | number>;
+    manualValues?: Map<string, string | number>;
+    visible?: boolean;
+}
+
+
+export enum PROPERTY_SELECTOR_SOURCE {
+    fix = 'Fix',
+    provided = 'Provided',
+    generated = 'Generated',
+    manual = 'Manual',
+    interpolated = 'Interpolated',
+    metric_on_field = 'Metric on field',
+    heatmap_density = 'Density'
+  }
