+++ projects/arlas-components/src/lib/components/results/utils/enumerations/thumbnailFitEnum.ts
