+++ src/index.ts
@@ -17,48 +17,3 @@
  * under the License.
  */
 
-export { ColorGeneratorLoader, AwcColorGeneratorLoader } from './components/componentsUtils';
-export { MapLayers, BasemapStyle, PaintColor, FillStroke, LayerMetadata, ARLAS_ID,
-  SCROLLABLE_ARLAS_ID, FILLSTROKE_LAYER_PREFIX, HOVER_LAYER_PREFIX, SELECT_LAYER_PREFIX,
-  ARLAS_VSET } from './components/mapgl/model/mapLayers';
-export { MapSource } from './components/mapgl/model/mapSource';
-export { MapExtend } from './components/mapgl/mapgl.component.util';
-export { PowerbarsComponent } from './components/powerbars/powerbars.component';
-export { PowerbarsModule } from './components/powerbars/powerbars.module';
-export { MetricComponent } from './components/metric/metric.component';
-export { MetricModule } from './components/metric/metric.module';
-export { ChartType, DataType, Position, SimpleNode, TreeNode, SwimlaneMode, HistogramTooltip } from 'arlas-d3';
-export { SortEnum } from './components/results/utils/enumerations/sortEnum';
-export { ModeEnum } from './components/results/utils/enumerations/modeEnum';
-export { ThumbnailFitEnum } from './components/results/utils/enumerations/thumbnailFitEnum';
-export { CellBackgroundStyleEnum } from './components/results/utils/enumerations/cellBackgroundStyleEnum';
-export { DetailedDataRetriever } from './components/results/utils/detailed-data-retriever';
-export { Action, ElementIdentifier, FieldsConfiguration, ResultListOptions } from './components/results/utils/results.utils';
-export { Item } from './components/results/model/item';
-export { Column } from './components/results/model/column';
-export { HistogramComponent } from './components/histogram/histogram.component';
-export { HistogramModule } from './components/histogram/histogram.module';
-export { DonutComponent } from './components/donut/donut.component';
-export { DonutModule } from './components/donut/donut.module';
-export { MapglComponent, VisualisationSetConfig, IconConfig, CROSS_LAYER_PREFIX } from './components/mapgl/mapgl.component';
-export { MapglModule } from './components/mapgl/mapgl.module';
-export { ResultListComponent } from './components/results/result-list/result-list.component';
-export { ResultItemComponent } from './components/results/result-item/result-item.component';
-export { ResultDetailedItemComponent } from './components/results/result-detailed-item/result-detailed-item.component';
-export { ResultsModule } from './components/results/results.module';
-export { ColorGeneratorModule } from './services/color.generator.module';
-export { ArlasColorService } from './services/color.generator.service';
-export { MapglImportModule } from './components/mapgl-import/mapgl-import.module';
-export { MapglImportComponent } from './components/mapgl-import/mapgl-import.component';
-export { MapglLegendModule } from './components/mapgl-legend/mapgl-legend.module';
-export { LayerIdToName } from './components/mapgl-legend/layer-name.pipe';
-export { MapglLegendComponent } from './components/mapgl-legend/mapgl-legend.component';
-export { MapglLayerIconModule } from './components/mapgl-layer-icon/mapgl-layer-icon.module';
-export { MapglLayerIconComponent } from './components/mapgl-layer-icon/mapgl-layer-icon.component';
-export { FormatNumberModule } from './pipes/format-number/format-number.module';
-export { FormatNumberPipe } from './pipes/format-number/format-number.pipe';
-export { MapglSettingsModule } from './components/mapgl-settings/mapgl-settings.module';
-export { MapglSettingsComponent, MapSettingsService, GeoQuery,
-  GeometrySelectModel, OperationSelectModel } from './components/mapgl-settings/mapgl-settings.component';
-export { WmtsLayerManagerModule } from './components/wmts-layer-manager/wmts-layer-manager.module';
-export { WmtsLayerManagerComponent } from './components/wmts-layer-manager/wmts-layer-manager.component';
